add student logic

+++ src/models/Person.java
@@ -2,12 +2,12 @@ package models;
 
 public class Person {
     private String name;
-    private short age;
+    private int age;
     private static int studentsNumber = 0;
 
     public Person() { studentsNumber++; }
 
-    public Person(String name, short age) {
+    public Person(String name, int age) {
         this.name = name;
         this.age = age;
     }
@@ -20,11 +20,11 @@ public class Person {
         this.name = name;
     }
 
-    public short getAge() {
+    public int getAge() {
         return age;
     }
 
-    public void setAge(short age) {
+    public void setAge(int age) {
         this.age = age;
     }
